Add redux to the package.json and refactor the organization of the files
