+++ package.json
@@ -13,7 +13,9 @@
     "prop-types": "^15.5.8",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
+    "react-redux": "^5.0.5",
     "react-router": "^3.0.1",
+    "redux": "^3.6.0",
     "sass-loader": "^6.0.3",
     "susy": "^2.2.12"
   },
@@ -50,5 +52,9 @@
     ]
   },
   "author": "Denis Žoljom",
-  "license": "GPL"
+  "license" : "(ISC OR GPL-3.0)",
+  "repository" : {
+    "type" : "git",
+    "url" : "https://github.com/dingo-d/wp-api.git"
+  }
 }
